Refactroing state management using Context API
