+++ src/components/Progress.tsx
@@ -1,14 +1,9 @@
 import { FC } from "react";
+import { useQuiz } from "../contexts/QuizContext";
 
-type ProgressProps = {
-  index: number;
-  numQuestions: number;
-  points: number;
-  maxPossiblePoints: number;
-  answer: number | null;
-};
+export const Progress: FC = () => {
+  const {index, numQuestions, points, maxPossiblePoints, answer} = useQuiz();
 
-export const Progress: FC<ProgressProps> = ({index, numQuestions, points, maxPossiblePoints, answer}) => {
   return (
     <header className="progress">
       <progress max={numQuestions} value={index + Number(answer !== null)} />
