+++ src/index.tsx
@@ -2,13 +2,16 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './components/App';
+import { QuizContextProvider } from './contexts/QuizContext';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
 );
 root.render(
   <React.StrictMode>
-    <App />
+    <QuizContextProvider>
+      <App />
+    </QuizContextProvider>
   </React.StrictMode>
 );
